+++ apps/its-battistar-be/scripts/setup-environment.ts
@@ -11,10 +11,6 @@ import { seedDB } from './seed-mongo';
 async function setupEnvironment() {
   try {
     const started = await setup();
-    if (!started) {
-      console.warn('❌ Did not start the services, skipping seeding');
-      return;
-    }
     if (!started.commonStarted) {
       console.warn('❌ Did not start the common services, skipping seeding');
       return;
@@ -25,16 +21,16 @@ async function setupEnvironment() {
     await seedDB();
 
     // await teardown();
-    mongoose.disconnect();
+    await mongoose.disconnect();
     process.exit(0);
   } catch (e) {
     console.error(e);
-    mongoose.disconnect();
+    await mongoose.disconnect();
     process.exit(1);
   }
 }
 
-setupEnvironment().catch((e) => {
+setupEnvironment().catch((e: unknown) => {
   console.error(e);
   process.exit(1);
 });