+++ apps/its-battistar-be/src/utils/timer.ts
@@ -1,7 +1,7 @@
-export const exitTimer = async function (interval: number) {
+export const exitTimer = async function (interval: number): Promise<void> {
   return new Promise<void>((resolve) => {
     const fn = () => {
-      // eslint-disable-next-line n/no-process-exit, unicorn/no-process-exit, no-magic-numbers
+      // eslint-disable-next-line n/no-process-exit, no-magic-numbers
       resolve(process.exit(0));
     };
     setInterval(fn, interval);