+++ apps/its-battistar/eslint.config.js
@@ -1,5 +1,9 @@
-const { FlatCompat } = require('@eslint/eslintrc');
+/* eslint-disable unicorn/prefer-module */
 const baseConfig = require('../../eslint.config.js');
+const globals = require('globals');
+const tseslint = require('typescript-eslint');
+
+const { FlatCompat } = require('@eslint/eslintrc');
 const js = require('@eslint/js');
 
 const compat = new FlatCompat({
@@ -7,24 +11,46 @@ const compat = new FlatCompat({
   recommendedConfig: js.configs.recommended,
 });
 
-module.exports = [
+module.exports = tseslint.config(
   ...baseConfig,
+
+  {
+    ignores: ['!**/*', 'storybook-static', 'node_modules', 'dist'],
+  },
+
   {
     languageOptions: {
       parserOptions: {
-        project: [
-          'tsconfig.*?.json',
-          // '.storybook/tsconfig.json'
-        ],
+        project: ['tsconfig.*?.json', '.storybook/tsconfig.json'],
+        tsconfigRootDir: __dirname,
+      },
+      globals: {
+        ...globals.browser,
+        ...globals.serviceworker,
       },
     },
   },
+
+  {
+    rules: {},
+  },
+
+  // FIXME: old compatibility mode
+  // do these even work?
   ...compat
     .config({
       extends: [
         'plugin:@nx/angular',
+        'plugin:@angular-eslint/recommended',
         'plugin:@angular-eslint/template/process-inline-templates',
+        'plugin:angular/johnpapa',
       ],
+      // why do i need to specify this in the overrides too?
+      parserOptions: {
+        // project: ['./tsconfig.*?.json'],
+        tsconfigRootDir: __dirname,
+        project: ['tsconfig.*?.json', '.storybook/tsconfig.json'],
+      },
     })
     .map((config) => ({
       ...config,
@@ -48,12 +74,16 @@ module.exports = [
         ],
       },
     })),
+
+  // FIXME: doesn't lint html files
   ...compat
-    .config({ extends: ['plugin:@nx/angular-template'] })
+    .config({
+      parser: '@angular-eslint/template-parser',
+      extends: ['plugin:@nx/angular-template'],
+    })
     .map((config) => ({
       ...config,
       files: ['**/*.html'],
       rules: {},
-    })),
-  { ignores: ['storybook-static'] },
-];
+    }))
+);