+++ apps/its-battistar/src/app/components/sidebar/sidebar.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
 import { IMenuItem } from '../../pages/dashboard/dashboard.service';
@@ -16,6 +16,7 @@ export class SidebarComponent {
   title = 'Sidebar';
 
   // TODO: move to signal input
-  @Input()
-  menuItems: IMenuItem[] = [];
+  // @Input()
+  // menuItems: IMenuItem[] = [];
+  menuItems = input<IMenuItem[]>();
 }