+++ eslint.config.js
@@ -1,40 +1,110 @@
+/* eslint-disable unicorn/prefer-module */
+/* eslint-disable no-magic-numbers */
 const globals = require('globals');
 const { FlatCompat } = require('@eslint/eslintrc');
 const nodePlugin = require('eslint-plugin-n');
 const nxEslintPlugin = require('@nx/eslint-plugin');
 const eslintPluginSimpleImportSort = require('eslint-plugin-simple-import-sort');
-const eslintPluginImport = require('eslint-plugin-import');
-const js = require('@eslint/js');
+const tseslint = require('typescript-eslint');
+const eslintPluginUnicorn = require('eslint-plugin-unicorn');
+const jsoncParser = require('jsonc-eslint-parser');
+const tsParser = require('@typescript-eslint/parser');
+const eslint = require('@eslint/js');
+
+// const eslintrc = require('@eslint/eslintrc');
+// const eslintPluginImport = require('eslint-plugin-import');
 
 const compat = new FlatCompat({
   baseDirectory: __dirname,
-  recommendedConfig: js.configs.recommended,
+  recommendedConfig: eslint.configs.recommended,
 });
 
-module.exports = [
+// TODO: fix nx lint
+
+// tseslint.config is an utility function provided by typescript-eslint
+// to provide type safety and intellisense to the configuration
+module.exports = tseslint.config(
   {
+    // TODO: fix ignores chains
+    // must be on its own for glob pattern to work
+    ignores: ['!**/*', 'node_modules', 'dist', 'tmp', '.angular'],
+  },
+
+  {
+    // root level config
     languageOptions: {
+      parser: tsParser,
       globals: {
-        ...globals.browser,
+        // ...globals.browser,
         ...globals.node,
-        ...globals.worker,
+        // ...globals.worker,
         ...globals.es2021,
-        ...globals.jest,
+        // ...eslintrc.Legacy.environments.get('es2024'),
+        // ...globals.jest,
       },
     },
-  },
-  {
     plugins: {
+      // provide the rules at the root level, activate the recommended config in express only
+      // or they won't be available globally
+      // this won't activate them
+      // you either need to extend or enable rules manually
       '@nx': nxEslintPlugin,
 
       'simple-import-sort': eslintPluginSimpleImportSort,
 
-      import: eslintPluginImport,
+      // BUG: esling-plugin-import doesn't support flat config
+      // import: eslintPluginImport,
 
-      // provide the rules at the root level, activate the recommended config in express only
+      unicorn: eslintPluginUnicorn,
+
       n: nodePlugin,
     },
   },
+
+  // global configs, any file type
+  eslint.configs.recommended,
+
+  // global rules, any file type
+  {
+    rules: {
+      'no-magic-numbers': 'error',
+      complexity: ['error', 20],
+      'prefer-const': 'warn',
+      eqeqeq: 'error',
+      'no-console': 'warn',
+
+      // using tsconfig noImplicitReturns
+      'consistent-return': 'off', // force explicit return, prevents bugs
+      'no-useless-return': 'off',
+
+      'simple-import-sort/imports': 'error',
+      'simple-import-sort/exports': 'error',
+
+      'sort-imports': 'off',
+
+      'unicorn/no-abusive-eslint-disable': 'off',
+      'unicorn/no-null': 'off',
+      'unicorn/no-array-for-each': 'warn',
+      'unicorn/consistent-function-scoping': 'warn',
+      'unicorn/prefer-top-level-await': 'warn',
+      'unicorn/prevent-abbreviations': 'off',
+      'unicorn/prefer-module': 'warn',
+    },
+  },
+
+  // filetype specific rules and configs
+  {
+    files: ['*.json'],
+    languageOptions: {
+      parser: jsoncParser,
+    },
+    rules: {
+      // FIXME: could not find plugin nx
+      // '@nx/dependency-checks': 'error', // TODO:
+    },
+  },
+
+  // all files
   {
     files: ['**/*.ts', '**/*.tsx', '**/*.js', '**/*.jsx'],
     rules: {
@@ -53,42 +123,67 @@ module.exports = [
       ],
     },
   },
+
+  // NOTE: this is how you map a config to a filetype
+  // alternatively you can use the extend utility provided by typescript-eslint
+
+  // typescript files
+  // ...[
+  //   tseslint.configs.eslintRecommended, // disable eslint rules already covered by typescript-eslint
+  //   ...tseslint.configs.strictTypeChecked,
+  //   ...tseslint.configs.stylisticTypeChecked,
+  // ].map((conf) => ({
+  //   ...conf,
+  //   files: ['**/*.ts', '**/*.tsx'],
+  //   rules: {
+  //     '@typescript-eslint/explicit-module-boundary-types': ['error'],
+  //   },
+  // })),
+
+  // alternative way as the one above
+  {
+    files: ['**/*.ts', '**/*.tsx'],
+    // extends is an utiliy function provided by typescript-eslint
+    extends: [
+      tseslint.configs.eslintRecommended, // disable eslint rules already covered by typescript-eslint
+      ...tseslint.configs.strictTypeChecked,
+      ...tseslint.configs.stylisticTypeChecked,
+    ],
+    rules: {
+      '@typescript-eslint/explicit-module-boundary-types': ['error'],
+    },
+  },
+
+  // test files
+  {
+    files: ['**/*.spec.ts', '**/*.spec.tsx', '**/*.spec.js', '**/*.spec.jsx'],
+    languageOptions: {
+      globals: {
+        ...globals.jest,
+      },
+    },
+    rules: {},
+  },
+
   // FIXME: compatility mode after nx migration
+  // do they override previous rules?
+  // do these even work??
+
+  // all files
   ...compat
     .config({
-      plugins: ['simple-import-sort', '@typescript-eslint'],
+      plugins: ['sonarjs'],
       extends: [
-        'plugin:@nx/typescript',
-        'eslint:recommended',
-        'plugin:@typescript-eslint/eslint-recommended',
-        'plugin:@typescript-eslint/strict-type-checked',
-        'plugin:@typescript-eslint/stylistic-type-checked',
-        'plugin:unicorn/recommended',
-        'plugin:import/recommended',
+        'plugin:@nx/javascript',
+        'plugin:import/recommended', // TODO:
+        'plugin:sonarjs/recommended',
       ],
     })
     .map((config) => ({
       ...config,
-      files: ['**/*.ts', '**/*.tsx'],
+      // files: ['**/*.js', '**/*.jsx', '**/*.ts', '**/*.tsx'],
       rules: {
-        'no-magic-numbers': 'error',
-        'sort-imports': 'off',
-        'simple-import-sort/imports': 'error',
-        'simple-import-sort/exports': 'error',
-        'unicorn/no-abusive-eslint-disable': 'off',
         'import/no-unresolved': 'off',
-        'unicorn/no-null': 'off',
-        'unicorn/no-array-for-each': 'warn',
-        'unicorn/consistent-function-scoping': 'warn',
-        'unicorn/prefer-top-level-await': 'warn',
-        'unicorn/prevent-abbreviations': 'off',
-        'unicorn/prefer-module': 'warn',
-        complexity: ['error', 20],
-        'consistent-return': 'warn',
-        'no-useless-return': 'warn',
-        'prefer-const': 'warn',
-        eqeqeq: 'error',
-        'no-console': 'warn',
         'import/no-extraneous-dependencies': 'off',
         'import/first': 'error',
         // 'import/newline-after-import': 'error', // BUG: new config
@@ -100,14 +195,38 @@ module.exports = [
         'import/namespace': 'off',
       },
     })),
-  ...compat.config({ extends: ['plugin:@nx/javascript'] }).map((config) => ({
-    ...config,
-    files: ['**/*.js', '**/*.jsx'],
-    rules: {},
-  })),
-  ...compat.config({ env: { jest: true } }).map((config) => ({
-    ...config,
-    files: ['**/*.spec.ts', '**/*.spec.tsx', '**/*.spec.js', '**/*.spec.jsx'],
-    rules: {},
-  })),
-];
+
+  // typescript files
+  ...compat
+    .config({
+      //     plugins: [
+      //       // '@typescript-eslint'
+      //     ],
+      extends: [
+        'plugin:@nx/typescript', //TODO:
+      ],
+    })
+    .map((config) => ({
+      ...config,
+      files: ['**/*.ts', '**/*.tsx'],
+      rules: {},
+    }))
+
+  // javascript files only
+  // ...compat
+  //   .config({
+  //     extends: [ ],
+  //   })
+  //   .map((config) => ({
+  //     ...config,
+  //     files: ['**/*.js', '**/*.jsx'],
+  //     rules: {},
+  //   }))
+
+  // spec files
+  // ...compat.config({ env: { jest: true } }).map((config) => ({
+  //   ...config,
+  //   files: ['**/*.spec.ts', '**/*.spec.tsx', '**/*.spec.js', '**/*.spec.jsx'],
+  //   rules: {},
+  // })),
+);