+++ libs/shared-types/eslint.config.js
@@ -1,12 +1,26 @@
+/* eslint-disable unicorn/prefer-module */
 const globals = require('globals');
 const baseConfig = require('../../eslint.config.js');
+const tseslint = require('typescript-eslint');
 
-module.exports = [
+module.exports = tseslint.config(
   ...baseConfig,
+
+  {
+    ignores: ['!**/*', 'node_modules', 'dist'],
+  },
+
   {
     languageOptions: {
-      parserOptions: { project: ['tsconfig.*?.json'] },
-      globals: { ...globals.node, ...globals.es2024 },
+      parserOptions: {
+        project: ['tsconfig.*?.json'],
+        tsconfigRootDir: __dirname,
+      },
+      globals: {
+        ...globals.node,
+        ...globals.es2021,
+        ...globals.browser,
+      },
     },
   },
   // {
@@ -21,5 +35,5 @@ module.exports = [
   {
     files: ['**/*.js', '**/*.jsx'],
     rules: {},
-  },
-];
+  }
+);